♻️ refactor(pong): simplify pong service and add timestamps
refactor the PongReturnService to use a single StartAsync method, improving clarity and removing redundant code
add timestamps to the console output for better event tracking and debugging
remove unnecessary dependency injection and command classes, simplifying the overall structure
