+++ src/ZtrConnectionTester.Console/Commands/PingPongPair/PongCommand/PongReturnService.cs
@@ -1,31 +1,28 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Spectre.Console;
-using Spectre.Console.Cli;
+﻿using Spectre.Console;
 using System;
 using System.IO;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using ZtrConnectionTester.Console.Commands.Base;
 using ZtrConnectionTester.Console.Commands.PingPongPair.Commons;
 
 namespace ZtrConnectionTester.Console.Commands.PingPongPair.PongCommand;
 
 public interface IPongReturnService
 {
-    Task ReceivePackageAndReturnItAsync(Stream stream, CancellationToken cancellationToken);
+    Task StartAsync(Stream stream, CancellationToken cancellationToken);
 }
 public class PongReturnService(IAnsiConsole console) : IPongReturnService
 {
-    public async Task StartPongTask(Stream stream, CancellationToken cancellationToken)
+    public async Task StartAsync(Stream stream, CancellationToken cancellationToken)
     {
         do
         {
             await ReceivePackageAndReturnItAsync(stream, cancellationToken);
         } while (!cancellationToken.IsCancellationRequested);
     }
-    public async Task ReceivePackageAndReturnItAsync(Stream stream, CancellationToken cancellationToken)
+
+    private async Task ReceivePackageAndReturnItAsync(Stream stream, CancellationToken cancellationToken)
     {
         try
         {
@@ -41,7 +38,7 @@ public class PongReturnService(IAnsiConsole console) : IPongReturnService
                 using var writer = new StreamWriter(stream, Encoding.ASCII, leaveOpen: true);
                 await writer.WriteAsync("PONG\n");
                 await writer.FlushAsync(cancellationToken);
-                console.WriteLine("PONG sent back!");
+                console.WriteLine($"[{DateTime.Now:HH:mm:ss.fff}] PONG sent back!");
             }
             else
             {
@@ -57,42 +54,5 @@ public class PongReturnService(IAnsiConsole console) : IPongReturnService
         {
             // Nothing interesting
         }
-    }
-}
-
-public static class SerialPongDependencyInjection
-{
-    public static IServiceCollection RegisterSerialPongCommand(this IServiceCollection services)
-    {
-        services.AddSingleton<ISerialPortEnumerator, DefaultSerialPortEnumerator>();
-        services.AddTransient<IPongReturnService, PongReturnService>();
-
-        return services;
-    }
-}
-
-public class SerialPongSettings(ISerialPortEnumerator portEnumerator) : BaseSerialPingSettings(portEnumerator)
-{
-}
-
-public class SerialPongCommand(
-    IAnsiConsole console,
-    IPongReturnService pongReturnService,
-    ILogger<SerialPongCommand> logger) : CancellableAsyncCommand<SerialPongSettings>
-{
-    private readonly ILogger<SerialPongCommand> _logger = logger;
-
-    public override async Task<int> ExecuteAsync(CommandContext context, SerialPongSettings settings,
-        CancellationToken cancellationToken)
-    {
-        console.MarkupLine("[yellow]Press Ctrl+C to stop[/]");
-        using var serialPort = settings.OpenPortWithVisualSummary(console);
-
-        do
-        {
-            await pongReturnService.ReceivePackageAndReturnItAsync(serialPort.BaseStream, cancellationToken);
-        } while (!cancellationToken.IsCancellationRequested);
-
-        return 0;
     }
 }
