+++ src/ZtrConnectionTester.Console/Commands/PingPongPair/PongCommand/SerialPongSettings.cs
@@ -1,0 +1,7 @@
+using ZtrConnectionTester.Console.Commands.PingPongPair.Commons;
+
+namespace ZtrConnectionTester.Console.Commands.PingPongPair.PongCommand;
+
+public class SerialPongSettings(ISerialPortEnumerator portEnumerator) : BaseSerialPingSettings(portEnumerator)
+{
+}
